delete old span extract parser; move extract_task_dependencies into separate file
