+++ workspace/pynb_dag_runner/tests/core/test_dag_runner.py
@@ -20,7 +20,7 @@ from pynb_dag_runner.opentelemetry_helpers import (
     Spans,
     SpanRecorder,
 )
-from pynb_dag_runner.tasks.extract import extract_task_dependencies
+from pynb_dag_runner.opentelemetry_task_span_parser import extract_task_dependencies
 from pynb_dag_runner.helpers import A, one
 
 import opentelemetry as otel
